Update animated_sidebar.rs
The node height is at the beginning of the animation a few frames 0. This leads to flickering when moving up because the offset is 0 an not the page height. This is an easy fix but maybe it would be better to fix the node_size because this is very unintuitive.